Re-login when getting a 404 on /forgot route

+++ routes/forgot-password.js
@@ -94,12 +94,27 @@ function token (req, res) {
                 , password: newPass
                 , mustChangePass: true
                 }
+    var didReLogin = false
 
     req.log.warn('About to change password', { name: name })
-    couch.changePass(newAuth, function (er, cr, data) {
+    couch.changePass(newAuth, function CP (er, cr, data) {
+      if (cr && cr.statusCode === 404 && !didReLogin) {
+        // probably the admin session expired.
+        // try to re-login
+        return couch.tokenGet(function (er2, tok) {
+          if (er2 || !couch.valid(tok)) {
+            er = er || er2
+            return res.error(500, er, 'admin couchdb token failure')
+          }
+          didReLogin = true
+          couch.changePass(newAuth, CP)
+        })
+      }
+
       if (er || cr.statusCode >= 400) {
         return res.error(er, cr && cr.statusCode, data && data.error)
       }
+
       config.redis.client.del('pwrecover_' + hash, function () {})
       res.template('password-changed.ejs', { password: newPass, profile: null })
     })
@@ -148,6 +163,7 @@ function handle (req, res) {
                 couch.get(pu, PU)
               })
             }
+
             // actually doesn't exist.
             return res.error(404, er)
           })
